chore: Refactor shareTrackByID method in ShareService oc: 3850 (#36)
* chore: Refactor shareTrackByID method in ShareService oc: 3850

The shareTrackByID method in the ShareService has been refactored to simplify the code and improve readability. Instead of subscribing to an observable and transforming the socialShareText, it now directly shares the track by calling the share method with the trackId as a parameter. This change reduces unnecessary complexity and improves performance.

chore: Update hostToGeohubAppId in ConfigService

The hostToGeohubAppId object in the ConfigService has been updated to use the imported hostToGeohubAppId from 'wm-core/store/api/api.service' instead of the previous hardcoded values. This change allows for more flexibility and easier maintenance of the geohub app IDs based on the hostname.

* chore: Update defaultShareObj text in ShareService

The defaultShareObj text property in the ShareService has been updated to an empty string. This change ensures that the default text for sharing is empty, allowing users to input their own custom text when sharing content.

* chore: Remove empty text property in defaultShareObj

The code change removes the assignment of an empty string to the text property in the defaultShareObj object. This change was made in the ShareService file.