chore: add e2e test for non-accessible track
This commit adds a new e2e test for app 52 to verify that the non-accessibility message of the track is correctly displayed. The test selects the layer "Tracce per tests e2e" and the track "Traccia non accessibile", and checks that the non-accessibility message "Questo sentiero non è accessibile" is correctly displayed.