chore: Add e2e tests for sharing track and poi oc: 4927 (#86)
* chore: Add e2e tests for sharing track and poi oc: 4927

This commit adds end-to-end tests for sharing a track and a point of interest (poi). The tests ensure that the share button is correctly displayed when selecting a track or poi, and that clicking on the share button opens the share window. The test utils file is also added, which contains helper functions for navigating to the home page, opening layers, pois, and tracks.

* chore: Remove unnecessary wait in share-track-and-poi.cy.ts

The code changes remove the unnecessary wait(1000) statements in the 'Should open share window when click on share button' test case. This improves the efficiency of the test by eliminating unnecessary delays.

---------

Co-authored-by: peppedeka <peppedeka@gmail.com>