refactor(e2e): ♻️ update selector for map details title
Replaced the `ion-title` selector with a more specific `[e2e-map-details-title]` selector across multiple test files to improve robustness and consistency in element selection. This change affects the following test files:

- `ec-track-details.cy.ts`
- `releted-poi.cy.ts`
- `tracks-edge.cy.ts`
- `ugc-tracks-map-click.cy.ts`

The update ensures that the tests accurately target the intended elements, reducing the risk of false positives or negatives due to selector mismatches.