refactor(map): 🔄 replace ion-title with ion-label for map details oc:5739 (#142)
* refactor(map): 🔄 replace ion-title with ion-label for map details

Replaced `ion-title` with `ion-label` for both UGC POI properties and track details in the map page HTML. Updated corresponding SCSS to style the `ion-label`, ensuring text alignment, multi-line display with ellipsis, and consistent margins. This change improves UI consistency and fixes layout issues related to text overflow.

* refactor(e2e): ♻️ update selector for map details title

Replaced the `ion-title` selector with a more specific `[e2e-map-details-title]` selector across multiple test files to improve robustness and consistency in element selection. This change affects the following test files:

- `ec-track-details.cy.ts`
- `releted-poi.cy.ts`
- `tracks-edge.cy.ts`
- `ugc-tracks-map-click.cy.ts`

The update ensures that the tests accurately target the intended elements, reducing the risk of false positives or negatives due to selector mismatches.

* style(theme): 💄 update styles for wm-map-details headers

Updated the CSS selectors for `wm-map-details` headers to use `ion-label` instead of `ion-title`. This change includes adjustments to the margin and padding for the header containers to improve visual alignment and consistency.

- Replaced `ion-title` with `ion-label` in relevant CSS rules.
- Adjusted the margin and padding for `.webmapp-info-header-container > ion-label`.
- Ensured consistent styling for `wm-map-details` and POI properties.