refactor(e2e): ♻️ rename and enhance tiles download test
Renamed the test file from `tails-download.cy.ts` to `tiles-download.cy.ts` to correct a typo. Enhanced the test by adding zoom actions and a wait to ensure the map is properly rendered before interacting with it. Updated the test description to reflect the changes in functionality, specifically focusing on downloading tiles by bounding box.

fix(environment): 🐛 update appId to 52

Updated the `appId` from 3 to 52 in the `environment.ts` file to reflect the correct application configuration.