chore(core): ✨ add icon loading action to app component
Added a new action to load icons in the app component. The `loadIcons` action is dispatched during the initialization phase of the application. This ensures that icons are loaded alongside other configurations and resources.

Additionally, removed unused selector `confPoisIcons` from the `map.page.ts` to clean up the code and improve maintainability.