chore(app): ✨ load icons on app initialization (#162)
Added the dispatch of `loadIcons()` action in the `AppComponent` constructor to ensure icons are loaded during the app's initialization phase.

feat(poi-properties): ✨ integrate readonly form in poi properties

Added a readonly form in the `poi-properties.component.html` to display POI form data if available. Updated the component to include necessary selectors and form group management.

Added a TODO in the HTML to consider merging `ugc-properties` with `poi-properties`.