Oc 6780 (#172)
* test: add e2e tests for home-result component oc: 6780

Introduces tests to verify the display and functionality
of the home-result component when layers are opened. Ensures
correct segment button is selected based on the layer type.

Relates to oc:6780

* chore: Update subproject references

Updates subproject commits to latest versions in shared modules.

* chore: Update subproject references oc: 6780

Updates the subproject commits for shared components.
Relates to branch oc_6780.